refactor(tests): consolidate integration tests from 8 to 6 files (#1014)
- Delete test_pr_workflow.py (4 mock-only unit tests, no real integration value)
- Merge test_inspect_pr_integration_advanced.py into test_inspect_pr_integration.py
- Merge test_minimal_profile.py into test_vibe_center_profile.py
- Result: 6 files, 34 tests (was 8 files, 38 tests), higher real integration ratio
